Fix bug with confusion matrix labels
If the validation set does not have any examples of a particular label from training, but an attempt is made to display all possible labels in the confusion matrix, there is an error unless the "labels" argument of plot_confusion_matrix is properly defined. This is now included with a comment near the top of the script.